Separated page from component
Page contains the buttons with the result component.
NotFound is a wrapper for Result
