+++ front/antd-demo/src/modules/wiki/pages/wiki-not-found/wiki-not-found.css
@@ -1,3 +1,8 @@
 .not-found-content{
     text-align: center;
+    margin-top: 10%;
+}
+
+.not-found-content button {
+    margin: 5px;
 }
